+++ packages/common/src/deprecated/hexToResourceId.ts
@@ -1,0 +1,4 @@
+import { hexToResource } from "../hexToResource";
+
+/** @deprecated use `hexToResource` instead */
+export const hexToResourceId = hexToResource;