+++ packages/common/src/index.ts
@@ -1,11 +1,22 @@
 export * from "./common";
 export * from "./createBurnerAccount";
-export * from "./createContract";
 export * from "./createNonceManager";
+export * from "./getContract";
 export * from "./getBurnerPrivateKey";
-export * from "./hexToResourceId";
+export * from "./getNonceManager";
+export * from "./getNonceManagerId";
+export * from "./hexToResource";
 export * from "./readHex";
-export * from "./resourceIdToHex";
+export * from "./resourceToHex";
 export * from "./resourceTypes";
+export * from "./sendTransaction";
 export * from "./spliceHex";
 export * from "./transportObserver";
+export * from "./writeContract";
+
+/** @deprecated use `getContract` instead */
+export { createContract } from "./deprecated/createContract";
+/** @deprecated use `resourceToHex` instead */
+export { resourceIdToHex } from "./deprecated/resourceIdToHex";
+/** @deprecated use `hexToResource` instead */
+export { hexToResourceId } from "./deprecated/hexToResourceId";