+++ packages/store-sync/src/postgres/index.ts
@@ -6,3 +6,4 @@ export * from "./schemaVersion";
 export * from "./postgresStorage";
 export * from "./setupTables";
 export * from "./syncToPostgres";
+export * from "./columnTypes";