+++ packages/block-logs-stream/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @latticexyz/block-logs-stream
 
+## 2.0.0-next.11
+
+### Patch Changes
+
+- f99e8898: Bump viem to 1.14.0 and abitype to 0.9.8
+- Updated dependencies [16b13ea8]
+- Updated dependencies [f99e8898]
+- Updated dependencies [d075f82f]
+  - @latticexyz/common@2.0.0-next.11
+
 ## 2.0.0-next.10
 
 ### Patch Changes